+++ changelog.md
@@ -1,3 +1,5 @@
+- add support for using remote data imediatly after activating the extension
+
 # 0.2.0
 - add support for specifying file resources as a hash so you can do
   ```ruby